requirements.txt
beautifulsoup4==4.12.3
dash==2.18.2
dash-bootstrap-components==1.6.0
numpy==2.2.4
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.40
torch==2.6.0
transformers==4.49.0
yfinance==0.2.54
